refactor: deactivate sales person when there's invoice, add set activate in update

+++ src/main/java/com/example/capstone2/Service/SalesPersonService.java
@@ -4,6 +4,7 @@ import com.example.capstone2.Api.Exception.ResourceNotFoundException;
 import com.example.capstone2.Api.Exception.SimpleException;
 import com.example.capstone2.DTO.UpdateSalesPersonDTO;
 import com.example.capstone2.Model.SalesPerson;
+import com.example.capstone2.Repository.SalesInvoiceRepository;
 import com.example.capstone2.Repository.SalesPersonRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,7 @@ import java.util.Objects;
 public class SalesPersonService {
 
     private final SalesPersonRepository salesPersonRepository;
+    private final SalesInvoiceRepository salesInvoiceRepository;
 
     public List<SalesPerson> findAll() {
         return salesPersonRepository.findAll();
@@ -40,6 +42,12 @@ public class SalesPersonService {
         }
 
         return salesPerson;
+    }
+
+    public void salesPersonExists(Integer id) throws SimpleException {
+        if(salesPersonRepository.findSalesPersonById(id) == null) {
+            throw new SimpleException("no sales person found with the sales person id you provided.");
+        }
     }
 
     public HashMap<String, Object> addSalesPerson(SalesPerson salesPerson) throws SimpleException {
@@ -73,6 +81,7 @@ public class SalesPersonService {
         salesPerson.setUsername(updateSalesPersonDTO.getUsername());
         salesPerson.setPassword(updateSalesPersonDTO.getPassword());
         salesPerson.setSalary(updateSalesPersonDTO.getSalary());
+        salesPerson.setActive(updateSalesPersonDTO.getActive());
 
 
         salesPersonRepository.save(salesPerson);
@@ -92,6 +101,19 @@ public class SalesPersonService {
             throw new ResourceNotFoundException("sales person");
         }
 
+        // you can not delete sales person if they have sales invoice
+        if(salesInvoiceRepository.atLeastOneSalesBySalesPersonId(id) != null) {
+            // instead deactivate their accounts.
+            salesPerson.setActive(false);
+            salesPersonRepository.save(salesPerson);
+
+            HashMap<String, Object> response = new HashMap<>();
+            response.put("message", "the sales person have been deactivated because there are invoices registered with their IDs.");
+            response.put("salesPerson", salesPerson);
+
+            return response;
+        }
+
         salesPersonRepository.deleteById(id);
 
         HashMap<String, Object> response = new HashMap<>();
@@ -99,5 +121,9 @@ public class SalesPersonService {
         response.put("salesPerson", salesPerson);
 
         return response;
+    }
+
+    public SalesPerson login(String username, String password) {
+        return salesPersonRepository.login(username, password);
     }
 }
